Improves UI redirection after successful authentication
This patch improves ui redirection by redirecting users to the same page
where they had attempted to login after successful authentication
