+++ ui/src/containers/Header/index.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { observer } from 'mobx-react';
 import { useHistory } from 'react-router-dom';
 import {
@@ -31,12 +31,56 @@ import { AUTH_BASE_URL, REDIRECT_URI } from '../../config/constants';
 import { IProvider } from '../../store/provider';
 import { titleCase } from '../../common/titlecase';
 import AlertDisplay from '../../components/AlertDisplay';
+import { AuthCodeProps, IError } from '../../store/auth';
 
 const Header: React.FC = observer(() => {
   const { user, providers } = useMst();
   const history = useHistory();
   const [isModalOpen, setIsModalOpen] = React.useState(false);
   const [disable, setDisable] = React.useState(false);
+
+  const detailPagePath = `detailPagePath`;
+
+  useEffect(() => {
+    // if user tries to authenticate from detail page
+    // Redirects to resource detail page
+    if (localStorage.getItem(detailPagePath) !== null && user.isAuthenticated === true) {
+      history.replace(`${localStorage.getItem(detailPagePath)}`);
+      localStorage.removeItem(detailPagePath);
+    }
+  }, [user.isAuthenticated]);
+
+  useEffect(() => {
+    // Parse the url and read the status and code params
+    const searchParams = new URLSearchParams(window.location.search);
+    const status = searchParams.get('status');
+    const code = searchParams.get('code');
+
+    // It checks status and code and then redirect to authentication
+    if (status === '200' && code !== null) {
+      const codeFinal: AuthCodeProps = {
+        code: code
+      };
+      user.authenticate(codeFinal);
+    }
+
+    // Strore the resource detail page path in the localStorage
+    if (history.location.pathname !== '/' && !user.isAuthenticated) {
+      localStorage.setItem(detailPagePath, history.location.pathname);
+    }
+
+    if (!user.isAuthenticated && status !== '200' && status !== null) {
+      // Wait to redirection of page and then update the store
+      setTimeout(() => {
+        const error: IError = {
+          status: Number(status),
+          serverMessage: 'Login Failed, Please Try To Login Again!',
+          customMessage: ''
+        };
+        user.setErrorMessage(error);
+      }, 1000);
+    }
+  }, [window.location.href]);
 
   const headerTools = (
     <PageHeaderTools>
