+++ ui/src/containers/ParseUrl/index.tsx
@@ -1,40 +1,13 @@
 import React from 'react';
-import { useHistory } from 'react-router-dom';
 import { useMst } from '../../store/root';
 import { Params } from '../../common/params';
-import { AuthCodeProps, IError } from '../../store/auth';
 
 const ParseUrl: React.FC = () => {
-  const { resources, user } = useMst();
-  const history = useHistory();
+  const { resources } = useMst();
 
   if (window.location.search) {
-    const searchParams: URLSearchParams = new URLSearchParams(window.location.search);
-    const status = searchParams.get('status');
-    const code = searchParams.get('code');
+    const searchParams = new URLSearchParams(window.location.search);
 
-    // It checks status and code and then redirect to authentication
-    if (status === '200' && code !== null) {
-      const codeFinal: AuthCodeProps = {
-        code: code
-      };
-      user.authenticate(codeFinal);
-      if (user.isAuthenticated) {
-        history.goBack();
-      }
-    }
-    // Display the alert message when status is not ok
-    else if (!user.isAuthenticated && status !== '200' && status !== null) {
-      // Wait to redirection of page and then update the store
-      setTimeout(() => {
-        const error: IError = {
-          status: Number(status),
-          serverMessage: 'Login Failed, Please Try To Login Again!',
-          customMessage: ''
-        };
-        user.setErrorMessage(error);
-      }, 1000);
-    }
     if (searchParams.has(Params.Query)) {
       resources.setSearch(searchParams.get(Params.Query) || '');
     }
